requirements.txt
networkx>=3.0
python-louvain>=0.16
scikit-learn>=1.0
numpy>=1.20
python-dotenv>=1.0.0
openai>=1.0.0
